+++ src/main/java/blackjack/card/Shape.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.card;
 
 public enum Shape {
     SPADE("스페이드"),
